Refactor claims architecture
After giving this some thought and discussing, decided to go back to a registry approach for claims for simplicity.

ClaimIssuerRegistry includes a `getClaim` method that is backwards compatible with ERC780. That is, if an issuer contract has not been set for the issuer, the contract delegates the request to ERC780.